+++ deps/vecmath/include/Matrix3f.h
@@ -12,10 +12,10 @@ class Matrix3f
 {
 public:
 	// Fill a 3x3 matrix with "fill", default to 0.
-	Matrix3f(double fill = 0.f);
-	Matrix3f(double m00, double m01, double m02,
-			 double m10, double m11, double m12,
-			 double m20, double m21, double m22);
+	Matrix3f(float fill = 0.f);
+	Matrix3f(float m00, float m01, float m02,
+			 float m10, float m11, float m12,
+			 float m20, float m21, float m22);
 
 	// setColumns = true ==> sets the columns of the matrix to be [v0 v1 v2]
 	// otherwise, sets the rows
@@ -25,8 +25,8 @@ public:
 	Matrix3f &operator=(const Matrix3f &rm); // assignment operator
 	// no destructor necessary
 
-	const double &operator()(int i, int j) const;
-	double &operator()(int i, int j);
+	const float &operator()(int i, int j) const;
+	float &operator()(int i, int j);
 
 	Vector3f getRow(int i) const;
 	void setRow(int i, const Vector3f &v);
@@ -42,35 +42,35 @@ public:
 	// starting with upper left corner at (i0, j0)
 	void setSubmatrix2x2(int i0, int j0, const Matrix2f &m);
 
-	double determinant() const;
-	Matrix3f inverse(bool *pbIsSingular = NULL, double epsilon = 0.f) const; // TODO: invert in place as well
+	float determinant() const;
+	Matrix3f inverse(bool *pbIsSingular = NULL, float epsilon = 0.f) const; // TODO: invert in place as well
 
 	void transpose();
 	Matrix3f transposed() const;
 
 	// ---- Utility ----
-	operator double *(); // automatic type conversion for GL
+	operator float *(); // automatic type conversion for GL
 	void print();
 
-	static double determinant3x3(double m00, double m01, double m02,
-								 double m10, double m11, double m12,
-								 double m20, double m21, double m22);
+	static float determinant3x3(float m00, float m01, float m02,
+								 float m10, float m11, float m12,
+								 float m20, float m21, float m22);
 
 	static Matrix3f ones();
 	static Matrix3f identity();
-	static Matrix3f rotateX(double radians);
-	static Matrix3f rotateY(double radians);
-	static Matrix3f rotateZ(double radians);
-	static Matrix3f scaling(double sx, double sy, double sz);
-	static Matrix3f uniformScaling(double s);
-	static Matrix3f rotation(const Vector3f &rDirection, double radians);
+	static Matrix3f rotateX(float radians);
+	static Matrix3f rotateY(float radians);
+	static Matrix3f rotateZ(float radians);
+	static Matrix3f scaling(float sx, float sy, float sz);
+	static Matrix3f uniformScaling(float s);
+	static Matrix3f rotation(const Vector3f &rDirection, float radians);
 
 	// Returns the rotation matrix represented by a unit quaternion
 	// if q is not normalized, it it normalized first
 	static Matrix3f rotation(const Quat4f &rq);
 
 private:
-	double m_elements[9];
+	float m_elements[9];
 };
 
 // Matrix-Vector multiplication
