+++ deps/vecmath/include/Vector2f.h
@@ -12,8 +12,8 @@ public:
 	static const Vector2f UP;
 	static const Vector2f RIGHT;
 
-	Vector2f(double f = 0.f);
-	Vector2f(double x, double y);
+	Vector2f(float f = 0.f);
+	Vector2f(float x, float y);
 
 	// copy constructors
 	Vector2f(const Vector2f &rv);
@@ -24,14 +24,14 @@ public:
 	// no destructor necessary
 
 	// returns the ith element
-	const double &operator[](int i) const;
-	double &operator[](int i);
+	const float &operator[](int i) const;
+	float &operator[](int i);
 
-	double &x();
-	double &y();
+	float &x();
+	float &y();
 
-	double x() const;
-	double y() const;
+	float x() const;
+	float y() const;
 
 	Vector2f xy() const;
 	Vector2f yx() const;
@@ -41,31 +41,31 @@ public:
 	// returns ( -y, x )
 	Vector2f normal() const;
 
-	double abs() const;
-	double absSquared() const;
+	float abs() const;
+	float absSquared() const;
 	void normalize();
 	Vector2f normalized() const;
 
 	void negate();
 
 	// ---- Utility ----
-	operator const double *() const; // automatic type conversion for OpenGL
-	operator double *();			 // automatic type conversion for OpenGL
+	operator const float *() const; // automatic type conversion for OpenGL
+	operator float *();			 // automatic type conversion for OpenGL
 	void print() const;
 
 	Vector2f &operator+=(const Vector2f &v);
 	Vector2f &operator-=(const Vector2f &v);
-	Vector2f &operator*=(double f);
+	Vector2f &operator*=(float f);
 
-	static double dot(const Vector2f &v0, const Vector2f &v1);
+	static float dot(const Vector2f &v0, const Vector2f &v1);
 
 	static Vector3f cross(const Vector2f &v0, const Vector2f &v1);
 
 	// returns v0 * ( 1 - alpha ) * v1 * alpha
-	static Vector2f lerp(const Vector2f &v0, const Vector2f &v1, double alpha);
+	static Vector2f lerp(const Vector2f &v0, const Vector2f &v1, float alpha);
 
 private:
-	double m_elements[2];
+	float m_elements[2];
 };
 
 // component-wise operators
@@ -78,9 +78,9 @@ Vector2f operator/(const Vector2f &v0, const Vector2f &v1);
 Vector2f operator-(const Vector2f &v);
 
 // multiply and divide by scalar
-Vector2f operator*(double f, const Vector2f &v);
-Vector2f operator*(const Vector2f &v, double f);
-Vector2f operator/(const Vector2f &v, double f);
+Vector2f operator*(float f, const Vector2f &v);
+Vector2f operator*(const Vector2f &v, float f);
+Vector2f operator/(const Vector2f &v, float f);
 
 bool operator==(const Vector2f &v0, const Vector2f &v1);
 bool operator!=(const Vector2f &v0, const Vector2f &v1);
