+++ deps/vecmath/src/Matrix2f.cpp
@@ -7,7 +7,7 @@
 
 #include "Vector2f.h"
 
-Matrix2f::Matrix2f(double fill)
+Matrix2f::Matrix2f(float fill)
 {
 	for (int i = 0; i < 4; ++i)
 	{
@@ -15,8 +15,8 @@ Matrix2f::Matrix2f(double fill)
 	}
 }
 
-Matrix2f::Matrix2f(double m00, double m01,
-				   double m10, double m11)
+Matrix2f::Matrix2f(float m00, float m01,
+				   float m10, float m11)
 {
 	m_elements[0] = m00;
 	m_elements[1] = m10;
@@ -41,24 +41,24 @@ Matrix2f::Matrix2f(const Vector2f &v0, const Vector2f &v1, bool setColumns)
 
 Matrix2f::Matrix2f(const Matrix2f &rm)
 {
-	memcpy(m_elements, rm.m_elements, 2 * sizeof(double));
+	memcpy(m_elements, rm.m_elements, 2 * sizeof(float));
 }
 
 Matrix2f &Matrix2f::operator=(const Matrix2f &rm)
 {
 	if (this != &rm)
 	{
-		memcpy(m_elements, rm.m_elements, 2 * sizeof(double));
+		memcpy(m_elements, rm.m_elements, 2 * sizeof(float));
 	}
 	return *this;
 }
 
-const double &Matrix2f::operator()(int i, int j) const
+const float &Matrix2f::operator()(int i, int j) const
 {
 	return m_elements[j * 2 + i];
 }
 
-double &Matrix2f::operator()(int i, int j)
+float &Matrix2f::operator()(int i, int j)
 {
 	return m_elements[j * 2 + i];
 }
@@ -93,16 +93,16 @@ void Matrix2f::setCol(int j, const Vector2f &v)
 	m_elements[colStart + 1] = v.y();
 }
 
-double Matrix2f::determinant()
+float Matrix2f::determinant()
 {
 	return Matrix2f::determinant2x2(
 		m_elements[0], m_elements[2],
 		m_elements[1], m_elements[3]);
 }
 
-Matrix2f Matrix2f::inverse(bool *pbIsSingular, double epsilon)
+Matrix2f Matrix2f::inverse(bool *pbIsSingular, float epsilon)
 {
-	double determinant = m_elements[0] * m_elements[3] - m_elements[2] * m_elements[1];
+	float determinant = m_elements[0] * m_elements[3] - m_elements[2] * m_elements[1];
 
 	bool isSingular = (fabs(determinant) < epsilon);
 	if (isSingular)
@@ -120,7 +120,7 @@ Matrix2f Matrix2f::inverse(bool *pbIsSingular, double epsilon)
 			*pbIsSingular = false;
 		}
 
-		double reciprocalDeterminant = 1.0f / determinant;
+		float reciprocalDeterminant = 1.0f / determinant;
 
 		return Matrix2f(
 			m_elements[3] * reciprocalDeterminant, -m_elements[2] * reciprocalDeterminant,
@@ -130,8 +130,8 @@ Matrix2f Matrix2f::inverse(bool *pbIsSingular, double epsilon)
 
 void Matrix2f::transpose()
 {
-	double m01 = (*this)(0, 1);
-	double m10 = (*this)(1, 0);
+	float m01 = (*this)(0, 1);
+	float m10 = (*this)(1, 0);
 
 	(*this)(0, 1) = m10;
 	(*this)(1, 0) = m01;
@@ -144,7 +144,7 @@ Matrix2f Matrix2f::transposed() const
 		(*this)(0, 1), (*this)(1, 1));
 }
 
-Matrix2f::operator double *()
+Matrix2f::operator float *()
 {
 	return m_elements;
 }
@@ -157,8 +157,8 @@ void Matrix2f::print()
 }
 
 // static
-double Matrix2f::determinant2x2(double m00, double m01,
-								double m10, double m11)
+float Matrix2f::determinant2x2(float m00, float m01,
+								float m10, float m11)
 {
 	return (m00 * m11 - m01 * m10);
 }
@@ -187,10 +187,10 @@ Matrix2f Matrix2f::identity()
 }
 
 // static
-Matrix2f Matrix2f::rotation(double degrees)
+Matrix2f Matrix2f::rotation(float degrees)
 {
-	double c = cos(degrees);
-	double s = sin(degrees);
+	float c = cos(degrees);
+	float s = sin(degrees);
 
 	return Matrix2f(
 		c, -s,
@@ -201,7 +201,7 @@ Matrix2f Matrix2f::rotation(double degrees)
 // Operators
 //////////////////////////////////////////////////////////////////////////
 
-Matrix2f operator*(double f, const Matrix2f &m)
+Matrix2f operator*(float f, const Matrix2f &m)
 {
 	Matrix2f output;
 
@@ -216,7 +216,7 @@ Matrix2f operator*(double f, const Matrix2f &m)
 	return output;
 }
 
-Matrix2f operator*(const Matrix2f &m, double f)
+Matrix2f operator*(const Matrix2f &m, float f)
 {
 	return f * m;
 }
