+++ deps/vecmath/src/Vector2f.cpp
@@ -19,13 +19,13 @@ const Vector2f Vector2f::UP = Vector2f(0, 1);
 // static
 const Vector2f Vector2f::RIGHT = Vector2f(1, 0);
 
-Vector2f::Vector2f(double f)
+Vector2f::Vector2f(float f)
 {
     m_elements[0] = f;
     m_elements[1] = f;
 }
 
-Vector2f::Vector2f(double x, double y)
+Vector2f::Vector2f(float x, float y)
 {
     m_elements[0] = x;
     m_elements[1] = y;
@@ -47,32 +47,32 @@ Vector2f &Vector2f::operator=(const Vector2f &rv)
     return *this;
 }
 
-const double &Vector2f::operator[](int i) const
+const float &Vector2f::operator[](int i) const
 {
     return m_elements[i];
 }
 
-double &Vector2f::operator[](int i)
+float &Vector2f::operator[](int i)
 {
     return m_elements[i];
 }
 
-double &Vector2f::x()
+float &Vector2f::x()
 {
     return m_elements[0];
 }
 
-double &Vector2f::y()
+float &Vector2f::y()
 {
     return m_elements[1];
 }
 
-double Vector2f::x() const
+float Vector2f::x() const
 {
     return m_elements[0];
 }
 
-double Vector2f::y() const
+float Vector2f::y() const
 {
     return m_elements[1];
 }
@@ -102,26 +102,26 @@ Vector2f Vector2f::normal() const
     return Vector2f(-m_elements[1], m_elements[0]);
 }
 
-double Vector2f::abs() const
+float Vector2f::abs() const
 {
     return sqrt(absSquared());
 }
 
-double Vector2f::absSquared() const
+float Vector2f::absSquared() const
 {
     return m_elements[0] * m_elements[0] + m_elements[1] * m_elements[1];
 }
 
 void Vector2f::normalize()
 {
-    double norm = abs();
+    float norm = abs();
     m_elements[0] /= norm;
     m_elements[1] /= norm;
 }
 
 Vector2f Vector2f::normalized() const
 {
-    double norm = abs();
+    float norm = abs();
     return Vector2f(m_elements[0] / norm, m_elements[1] / norm);
 }
 
@@ -131,12 +131,12 @@ void Vector2f::negate()
     m_elements[1] = -m_elements[1];
 }
 
-Vector2f::operator const double *() const
+Vector2f::operator const float *() const
 {
     return m_elements;
 }
 
-Vector2f::operator double *()
+Vector2f::operator float *()
 {
     return m_elements;
 }
@@ -161,7 +161,7 @@ Vector2f &Vector2f::operator-=(const Vector2f &v)
     return *this;
 }
 
-Vector2f &Vector2f::operator*=(double f)
+Vector2f &Vector2f::operator*=(float f)
 {
     m_elements[0] *= f;
     m_elements[1] *= f;
@@ -169,7 +169,7 @@ Vector2f &Vector2f::operator*=(double f)
 }
 
 // static
-double Vector2f::dot(const Vector2f &v0, const Vector2f &v1)
+float Vector2f::dot(const Vector2f &v0, const Vector2f &v1)
 {
     return v0[0] * v1[0] + v0[1] * v1[1];
 }
@@ -184,7 +184,7 @@ Vector3f Vector2f::cross(const Vector2f &v0, const Vector2f &v1)
 }
 
 // static
-Vector2f Vector2f::lerp(const Vector2f &v0, const Vector2f &v1, double alpha)
+Vector2f Vector2f::lerp(const Vector2f &v0, const Vector2f &v1, float alpha)
 {
     return alpha * (v1 - v0) + v0;
 }
@@ -218,17 +218,17 @@ Vector2f operator-(const Vector2f &v)
     return Vector2f(-v.x(), -v.y());
 }
 
-Vector2f operator*(double f, const Vector2f &v)
+Vector2f operator*(float f, const Vector2f &v)
 {
     return Vector2f(f * v.x(), f * v.y());
 }
 
-Vector2f operator*(const Vector2f &v, double f)
+Vector2f operator*(const Vector2f &v, float f)
 {
     return Vector2f(f * v.x(), f * v.y());
 }
 
-Vector2f operator/(const Vector2f &v, double f)
+Vector2f operator/(const Vector2f &v, float f)
 {
     return Vector2f(v.x() / f, v.y() / f);
 }
